requirements.txt
pandas
numpy
pyyaml
pytest